% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{actionButton}
\alias{actionButton}
\title{Action button}
\usage{
actionButton(inputId, label, icon = NULL, width = NULL, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button - a text label, but you could also use any other HTML, like an image.}

\item{icon}{An optional \link{icon} to appear on the button.}

\item{width}{The width of the input.}

\item{...}{Named attributes to be applied to the button or remaining parameters passed to uibutton, like \code{type}.}
}
\description{
Creates an action button  whose value is initially zero, and increments by one each time it is pressed.
}
