% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{addDeps}
\alias{addDeps}
\title{Internal function that adds dashboard dependencies to html.}
\usage{
addDeps(content)
}
\arguments{
\item{content}{Content to which dependencies need to be added.}
}
\value{
Content with appended dependencies.
}
\description{
Internal function that adds dashboard dependencies to html.
}

