% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite.R
\name{addsvg_sprite}
\alias{addsvg_sprite}
\title{Add svg to a sprite map}
\usage{
addsvg_sprite(newsvg, sprite, id = NULL)
}
\arguments{
\item{newsvg}{\code{xml_document} object with svg that will be added}

\item{sprite}{\code{xml_document} object with the sprite map}

\item{id}{sets the id of the new svg inside the sprite map, if is
NULL it looks for the id inside the \code{newsvg} object.}
}
\value{
invisible copy of new sprite map
}
\description{
Add svg to a sprite map
}
\details{
the sprite map object is modified in place
}
\examples{
library(xml2)
sprite <- xml_new_root("svg", "version" = "1.1", "xmlns" = "http://www.w3.org/2000/svg")
newsvg <- read_xml("<svg version='1.1' xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 68'>
<path d='M55.096 10.462c1.338-1.374.14-3.699-1.322-3.136-1.482.572-1'/>
   </svg>")
addsvg_sprite(newsvg, sprite, "new-icon")
}
