% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ALLOWED_BOX_SIDES_NONRIBBON}
\alias{ALLOWED_BOX_SIDES_NONRIBBON}
\title{Allowed box sizes}
\format{
An object of class \code{character} of length 6.
}
\usage{
ALLOWED_BOX_SIDES_NONRIBBON
}
\description{
Allowed box sizes
}
\keyword{datasets}
