% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ALLOWED_BOX_SIDES_RIBBON}
\alias{ALLOWED_BOX_SIDES_RIBBON}
\title{Allowed ribbon box sizes}
\format{
An object of class \code{character} of length 2.
}
\usage{
ALLOWED_BOX_SIDES_RIBBON
}
\description{
Allowed ribbon box sizes
}
\keyword{datasets}
