% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{appendDependencies}
\alias{appendDependencies}
\title{Internal function that adds an html dependency, without overwriting existing ones.}
\usage{
appendDependencies(content, newDependencies)
}
\arguments{
\item{content}{Content to which dependencies need to be added.}

\item{newDependencies}{Dependencies to be added.}
}
\value{
Content with appended dependencies.
}
\description{
Internal function that adds an html dependency, without overwriting existing ones.
}

