% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{apply_custom_styles_to_html_template}
\alias{apply_custom_styles_to_html_template}
\title{Format string template (that represents HTML template) with custom CSS styles.}
\usage{
apply_custom_styles_to_html_template(
  html_template = "",
  area_names = c(),
  container_style = "",
  area_styles = list()
)
}
\arguments{
\item{html_template}{string}

\item{area_names}{vector of strings}

\item{container_style}{string}

\item{area_styles}{list of strings}
}
\value{
string
}
\description{
Format string template (that represents HTML template) with custom CSS styles.
}
\details{
This is a helper function used in grid()
}
