% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{attach_rule}
\alias{attach_rule}
\title{Internal function that creates the rule for a specific setting or behavior of the modal.}
\usage{
attach_rule(id, behavior, target, value)
}
\arguments{
\item{id}{ID of the target modal.}

\item{behavior}{What behavior is beging set i. e. setting or attach events.}

\item{target}{First argument of the behavior. Usually a target or a setting name.}

\item{value}{Second argument of the behavior. usually an action or a setting value.}
}
\description{
Internal function that creates the rule for a specific setting or behavior of the modal.
}
\keyword{internal}
