% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_range.R
\name{update_calendar_range}
\alias{update_calendar_range}
\title{Update UI calendar range}
\usage{
update_calendar_range(
  session,
  input_id,
  start_value = NULL,
  end_value = NULL,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{input_id}{ID of the calendar range that will be updated}
}
\description{
This function updates the dates on a calendar range.
}
