% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar_range_single}
\alias{calendar_range_single}
\title{Defines calendar ranges}
\usage{
calendar_range_single(
  input_id,
  value = NULL,
  placeholder = NULL,
  type = "date",
  min = NA,
  max = NA,
  start_calendar_id = NULL,
  end_calendar_id = NULL
)
}
\arguments{
\item{input_id}{Input name. Reactive value is available under \code{input[[input_id]]}.}

\item{value}{Initial value of the numeric input.}

\item{placeholder}{Text visible in the input when nothing is inputted.}

\item{type}{Select from \code{'year'}, \code{'month'}, \code{'date'} and \code{'time'}}

\item{min}{Minimum allowed value.}

\item{max}{Maximum allowed value.}

\item{start_calendar_id}{id of the calendar that defines the range start.}

\item{end_calendar_id}{id of the calendar that defines the range end.}
}
\description{
Defines calendar ranges
}
