% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_extra_arguments}
\alias{check_extra_arguments}
\title{Check for extra arguments}
\usage{
check_extra_arguments(args)
}
\arguments{
\item{args}{list with extra arguments}
}
\description{
This throws warning if there're parameters not supported by semantic.
}
