% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{check_semantic_theme}
\alias{check_semantic_theme}
\title{Semantic theme path validator}
\usage{
check_semantic_theme(theme_css)
}
\arguments{
\item{theme_css}{it can be either NULL, character with css path, or theme name}
}
\value{
path to theme
}
\description{
Semantic theme path validator
}
\examples{
check_semantic_theme(NULL)
check_semantic_theme("darkly")
}
