% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_shiny_param}
\alias{check_shiny_param}
\title{Checks whether argument included as shiny exclusive parameter}
\usage{
check_shiny_param(name, func, ...)
}
\description{
Checks whether argument included as shiny exclusive parameter
}
