% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{checkbox_positions}
\alias{checkbox_positions}
\title{Checkbox positions}
\format{An object of class \code{character} of length 2.}
\usage{
checkbox_positions
}
\description{
Checkbox positions
}
\keyword{datasets}
