% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{checkbox_types}
\alias{checkbox_types}
\title{Checkbox types}
\format{An object of class \code{character} of length 4.}
\usage{
checkbox_types
}
\description{
Checkbox types
}
\keyword{datasets}
