% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkbox_ui}
\alias{checkbox_ui}
\title{'checkbox_ui' is deprecated. Use 'simple_checkbox' instead.}
\usage{
checkbox_ui(id, label, type = "", is_marked = TRUE, style = NULL)
}
\arguments{
\item{id}{Input name. Reactive value is available under input[[name]].}

\item{label}{Text to be displayed with checkbox.}

\item{type}{Type of checkbox. Please check \code{checkbox_types} for possibilities.}

\item{is_marked}{Defines if checkbox should be marked. Default TRUE.}

\item{style}{Style of the widget.}
}
\description{
This creates a checkbox using Semantic UI styles.
}
