% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{close_toast}
\alias{close_toast}
\title{Close Semantic UI toast}
\usage{
close_toast(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of the toast that will be displayed.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
This closes a displayed Semantic UI toast.
}
\seealso{
toast
}
