% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{COLOR_PALETTE}
\alias{COLOR_PALETTE}
\title{Semantic colors}
\format{
An object of class \code{character} of length 13.
}
\usage{
COLOR_PALETTE
}
\description{
https://github.com/Semantic-Org/Semantic-UI/blob/master/src/themes/default/globals/site.variables
}
\keyword{datasets}
