% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{counterbutton}
\alias{counterbutton}
\alias{counterButton}
\title{Counter Button}
\usage{
counterbutton(
  name,
  label = "",
  icon = NULL,
  value = 0,
  color = "",
  size = "",
  big_mark = " "
)

counterButton(
  name,
  label = "",
  icon = NULL,
  value = 0,
  color = "",
  size = "",
  big_mark = " "
)
}
\arguments{
\item{name}{The \code{input} slot that will be used to access the value.}

\item{label}{the content of the item to display}

\item{icon}{an optional \code{\link{uiicon}()} to appear on the button.}

\item{value}{initial rating value (integer)}

\item{color}{character with semantic color}

\item{big_mark}{big numbers separator}
}
\value{
counter button object
}
\description{
Creates a counter button whose value increments by one each time it is pressed.
}
\examples{
if (interactive()) {
library(shiny)
library(shiny.semantic)
ui <-semanticPage(
     counterbutton("counter", "My Counter Button",
                   icon = uiicon("world"),
                   size = "big", color = "purple")
 )
server <- function(input, output) {
 observeEvent(input$counter,{
   print("Counter", input$counter)
  })
 }
shinyApp(ui, server)
}
}
