% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{create_modal}
\alias{create_modal}
\title{Allows for the creation of modals in the server side without being tied to a specific HTML element.}
\usage{
create_modal(
  ui_modal,
  show = TRUE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{ui_modal}{HTML containing the modal.}

\item{show}{If the modal should only be created or open when called (open by default).}

\item{session}{Current session.}
}
\description{
Allows for the creation of modals in the server side without being tied to a specific HTML element.
}
