% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{define_selection_type}
\alias{define_selection_type}
\title{define_selection_type}
\usage{
define_selection_type(name, multiple)
}
\arguments{
\item{name}{- string with object name}

\item{multiple}{- logical}
}
\description{
define_selection_type
}
