% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{digits2words}
\alias{digits2words}
\title{Helper function that transforms digits to words}
\usage{
digits2words(number)
}
\arguments{
\item{number}{numeric digits from 1 to 10}
}
\value{
character with number word
}
\description{
Helper function that transforms digits to words
}
\keyword{internal}
