% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{display_grid}
\alias{display_grid}
\title{Display grid template in a browser for easy debugging}
\usage{
display_grid(grid_template)
}
\arguments{
\item{grid_template}{generated by grid_template() function}
}
\description{
Display grid template in a browser for easy debugging
}
\details{
Opens a browser and displays grid template with styled border to highlight existing areas.

Warning: CSS can't be displayed in RStudio viewer pane. CSS grid is supported only by modern browsers.
You can see list of supported browsers here: https://www.w3schools.com/css/css_grid.asp
}
