% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dropdown_choice}
\alias{dropdown_choice}
\title{Create Semantic UI Dropdown Choice Item}
\usage{
dropdown_choice(name, ..., link)
}
\arguments{
\item{name}{Name of the dropdown choice. This corresponds to 'value' argument for uidropdown.}

\item{...}{Content of the dropdown choice such as text, icons, labels.}

\item{link}{To what dropdown choice should be linked. If not specified, dropdown choice is not linked.}
}
\description{
This creates a dropdown item using Semantic UI.
}
