% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dropdown_divider}
\alias{dropdown_divider}
\title{Create Semantic UI Hedaer Item}
\usage{
dropdown_divider(...)
}
\arguments{
\item{...}{Other attributes of the divider such as style.}
}
\description{
This creates a dropdown header item using Semantic UI.
}
