% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dropdown_header}
\alias{dropdown_header}
\title{Create Semantic UI Hedaer Item}
\usage{
dropdown_header(...)
}
\arguments{
\item{...}{Content of the header: text, icons, etc.}
}
\description{
This creates a dropdown header item using Semantic UI.
}
