% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dropdown_menu}
\alias{dropdown_menu}
\title{Create Semantic UI Dropdown Menu}
\usage{
dropdown_menu(...)
}
\arguments{
\item{...}{Dropdown content. Use dropdown_choice to create dropdown item. Use dropdown_header to create dropdown header item.
Use dropdown_divider to create divider item.}
}
\description{
Create Semantic UI Dropdown Menu
}
