% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_icon_name}
\alias{extract_icon_name}
\title{Extract icon name}
\usage{
extract_icon_name(icon)
}
\arguments{
\item{icon}{icon object}
}
\value{
character with icon name
}
\description{
Extract icon name
}
\keyword{internal}
