% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{field_validation}
\alias{field_validation}
\alias{field_rule}
\title{Field Validation for Semantic UI}
\usage{
field_validation(id, ...)

field_rule(rule, prompt = NULL, value = NULL)
}
\arguments{
\item{id}{HTML id of the field to be validated}

\item{...}{A series of \code{field_rule}s that will be applied to the field}

\item{rule}{The type of rule to be applied. Valid rules are available in Details.}

\item{prompt}{Text to be displayed in the UI if the validation fails. Leave \code{NULL} if want to use default text.}

\item{value}{Certain fields require a value to check validation. Check Details if the \code{rule} requires a value.}
}
\description{
Field Validation for Semantic UI
}
\details{
The following \code{rule} are available:
\itemize{
}
}
\examples{
# E-mail validations
field_validation("email", field_rule("email"))

}
\seealso{
\code{form_validation}
}
