% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\docType{data}
\name{FOMANTIC_SIZE_LEVELS}
\alias{FOMANTIC_SIZE_LEVELS}
\title{allowed sizes}
\format{
An object of class \code{character} of length 7.
}
\usage{
FOMANTIC_SIZE_LEVELS
}
\description{
allowed sizes
}
\keyword{datasets}
