% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{form_button}
\alias{form_button}
\title{Fomantic UI Button}
\usage{
form_button(input_id, label, icon = NULL, width = NULL, class = NULL, ...)
}
\arguments{
\item{input_id}{The input slot that will be used to access the value}

\item{label}{The contents of the button, can either be character string or HTML tags}

\item{icon}{An optional \code{\link{icon}} to appear on the button}

\item{width}{Width of the input}

\item{class}{An optional attribute to be added to the button's class. If used
paramters like \code{color}, \code{size} are ignored.}

\item{...}{Named attributes to be applied to the button or remaining parameters passed to button, like \code{class}}
}
\value{
A \code{shiny.tag} that will show a submit button in the UI of a shiny application.
}
\description{
Creates a button specifically for Fomantic UI forms in order to check all inputs meet validation rules
}
\examples{
form_button("submit", "Submit")

}
\seealso{
\code{\link{form_validation}}, \code{\link{action_button}}
}
