% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_random_id}
\alias{generate_random_id}
\title{Some elements require input id, but this does not need to be
specified by the user. Thus we assign random value with prefix where needed.}
\usage{
generate_random_id(prefix, id_length = 20)
}
\arguments{
\item{prefix}{character with prefix add to id}

\item{id_length}{numeric with length of id (default 20)}
}
\description{
Some elements require input id, but this does not need to be
specified by the user. Thus we assign random value with prefix where needed.
}
