% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{get_cdn_path}
\alias{get_cdn_path}
\title{Get CDN path semantic dependencies}
\usage{
get_cdn_path()
}
\value{
CDN path of semantic dependencies
}
\description{
Internal function that return cdn path from options.
}
