% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{get_css_file}
\alias{get_css_file}
\title{Get css file}
\usage{
get_css_file(type, theme = NULL, minified = NULL)
}
\arguments{
\item{type}{define type of dependencies source}

\item{theme}{define theme}

\item{minified}{define if minified version should be used}
}
\value{
css file name
}
\description{
Get css file
}
\keyword{internal}
