% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{get_default_semantic_js}
\alias{get_default_semantic_js}
\title{Get default semantic js}
\usage{
get_default_semantic_js()
}
\value{
path to default js semantic file
}
\description{
Get default semantic js
}
