% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{get_default_semantic_theme}
\alias{get_default_semantic_theme}
\title{Get default semantic css}
\usage{
get_default_semantic_theme()
}
\value{
path to default css semantic file
}
\description{
Get default semantic css
}
