% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Add dashboard dependencies to html}
\usage{
get_dependencies(theme = NULL)
}
\arguments{
\item{theme}{define theme}
}
\value{
Content with appended dependencies.
}
\description{
Internal function that adds dashboard dependencies to html.
}
