% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{get_dependencies_path}
\alias{get_dependencies_path}
\title{Get dependencies path}
\usage{
get_dependencies_path()
}
\value{
path with css and js files
}
\description{
Get dependencies path
}
\keyword{internal}
