% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{get_numeric}
\alias{get_numeric}
\title{Extracts numeric values}
\usage{
get_numeric(value)
}
\arguments{
\item{value}{Value to be converted to numeric}
}
\value{
Numeric value
}
\description{
Extracts numeric values
}
\keyword{internal}
