% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{get_percent}
\alias{get_percent}
\title{Returns percentage value of width based on 12-units grid}
\usage{
get_percent(width)
}
\arguments{
\item{width}{value to be converted to percentage value}
}
\value{
percentage value of width based on 12-units grid
}
\description{
Returns percentage value of width based on 12-units grid
}
