% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_layout.R
\name{get_row}
\alias{get_row}
\title{Creates row wrapping specified elements}
\usage{
get_row(arg)
}
\arguments{
\item{arg}{Element to be wrapped in row container}
}
\value{
Row wrapping specified elements
}
\description{
Creates row wrapping specified elements
}
