% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{horizontal_menu}
\alias{horizontal_menu}
\title{Horizontal menu}
\usage{
horizontal_menu(menu_items, active_location = "", logo = NULL)
}
\arguments{
\item{menu_items}{data.frame with columns: "name", "link" and optional "icon"}

\item{active_location}{active location of the menu (should match
one from "link")}

\item{logo}{optional argument that displays logo on the left
of horizontal menu, can be character with image location, or shiny image object}
}
\value{
shiny div with horizontal menu
}
\description{
Renders UI with horizontal menu
}
\examples{
menu_content <- data.frame(
 name = paste("Menu", 1:3),
 link = paste("page", 1:3),
 icon = c("home", "car", 'tree')
)
if (interactive()){
  ui <- semanticPage(
   horizontal_menu(menu_content)
  )
  server <- function(input, output, session) {}
  shinyApp(ui, server)
}
}
