% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{hyperlink_item}
\alias{hyperlink_item}
\title{Create Semantic UI Menu Hyperlink Item}
\usage{
hyperlink_item(link = "", ..., item_feature = "", style = "")
}
\arguments{
\item{link}{To what the menu item should be linked. Default is "".}

\item{...}{Content of the hyperlink menu item: text, icons or labels to be displayed.}

\item{item_feature}{If required, add additional item feature like 'active', 'header', etc.}

\item{style}{Style of the item, e.g. "text-align: center".}
}
\description{
This creates a hyperlink menu item using Semantic UI
}
