% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{label}
\alias{label}
\title{Create HTML label tag}
\usage{
label(...)
}
\arguments{
\item{...}{Other arguments to be added as attributes of the
tag (e.g. style, class or childrens etc.)}
}
\description{
This creates a HTML label tag.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()){
  library(shiny.semantic)

  ui <- shinyUI(semanticPage(
    form(
      field(
        label("Name"),
        text_input("name", value = "", type = "text", placeholder = "Enter Name...")
      )
    )
  ))
  server <- shinyServer(function(input, output) {
  })

  shiny::shinyApp(ui, server)
}

}
