% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{list_element}
\alias{list_element}
\title{Helper function to render list element}
\usage{
list_element(header = NULL, description = NULL, icon = NULL)
}
\arguments{
\item{header}{character with header element}

\item{description}{character with content of the list}

\item{icon}{character with optional icon}
}
\description{
Helper function to render list element
}
