% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarLayout.R
\name{mainPanel}
\alias{mainPanel}
\title{Main Panel ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''}
\usage{
mainPanel(...)
}
\arguments{
\item{...}{Content of main panel}

\item{width}{Specified with of main panel}
}
\description{
Main Panel ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
}
