% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{menu_divider}
\alias{menu_divider}
\title{Create Semantic UI Divider Item}
\usage{
menu_divider(...)
}
\arguments{
\item{...}{Other attributes of the divider such as style.}
}
\description{
This creates a menu divider item using Semantic UI.
}
\seealso{
menu
}
