% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{menu_header}
\alias{menu_header}
\title{Create Semantic UI Header Item}
\usage{
menu_header(..., is_item = TRUE)
}
\arguments{
\item{...}{Content of the header: text, icons, etc.}

\item{is_item}{If TRUE created header is item of Semantic UI Menu.}
}
\description{
This creates a dropdown header item using Semantic UI.
}
