% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{menu_item}
\alias{menu_item}
\title{Create Semantic UI Menu Item}
\usage{
menu_item(..., item_feature = NULL, style = NULL, is_link = FALSE,
  link = NULL)
}
\arguments{
\item{...}{Content of the menu item: text, icons or labels to be displayed.}

\item{item_feature}{If required, add additional item feature like 'active', 'header', etc.}

\item{style}{Style of the item, e.g. "text-align: center".}

\item{is_link}{If TRUE menu item is created with 'a' tag - ortherwise with 'div' tag (default).}

\item{link}{If is_link is TRUE, determines to what the menu item should be linked. Ignored when is_link is FALSE.}
}
\description{
This creates a menu item using Semantic UI
}
