% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{multiple_checkbox}
\alias{multiple_checkbox}
\title{Create Semantic UI multiple checkbox}
\usage{
multiple_checkbox(
  name,
  label,
  choices,
  choices_value = choices,
  selected = NULL,
  position = "grouped",
  type = NULL,
  ...
)
}
\arguments{
\item{name}{Input name. Reactive value is available under input[[name]].}

\item{label}{Text to be displayed with checkbox.}

\item{choices}{List of values to show checkboxes for.
If elements of the list are named then that name rather than the value is displayed to the user.}

\item{selected}{The value that should be chosen initially.
If \code{NULL} the first one from \code{choices} is chosen.}

\item{position}{Specified checkmarks setup. Can be \code{grouped} or \code{inline}.}

\item{type}{Type of checkbox. Please check \code{\link{checkbox_types}} for possibilities.}

\item{...}{Other arguments to be added as attributes of the
tag (e.g. style, childrens etc.)}
}
\description{
This creates a multiple checkbox using Semantic UI styles.
}
