% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{multiple_radio}
\alias{multiple_radio}
\title{Create Semantic UI multiple radio buttons}
\usage{
multiple_radio(
  name,
  label,
  choices,
  choices_value = choices,
  selected = choices_value[1],
  position = "grouped",
  ...
)
}
\arguments{
\item{choices_value}{What reactive value should be used for corresponding
choice.}
}
\description{
This creates a multiple radiobox input using the Semantic UI style
}
