% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPage.R
\name{navbar_menu}
\alias{navbar_menu}
\title{Navebar Menu}
\usage{
navbar_menu(title, ..., menu_name = title, icon = NULL)
}
\arguments{
\item{title}{Display title for menu}

\item{...}{\code{\link{tab_panel}} elements to include in the page. Can also include strings as section headers,
or "----" as a horizontal separator.}

\item{menu_name}{The value that is linked to the \code{navbar_menu}}

\item{icon}{Optional icon to appear on the tab.
This attribute is only valid when using a \code{tab_panel} within a \code{\link{navbar_page}}.}
}
\description{

}
