% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_val}
\alias{parse_val}
\title{Parse the `shiny_input` value from JSON}
\usage{
parse_val(val)
}
\arguments{
\item{val}{value to get from JSON}
}
\value{
Value of type defined in `shiny_input`
}
\description{
Parse the `shiny_input` value from JSON
}
\keyword{internal}
