% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{remove_modal}
\alias{remove_modal}
\title{Hide Semantic UI modal}
\usage{
remove_modal(selector)
}
\arguments{
\item{target}{Javascript selector for the modal that will be hidden.}
}
\description{
This hides a displayed Semantic UI modal.
}
