% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite.R
\name{remove_svg_sprite}
\alias{remove_svg_sprite}
\title{Remove svg from a sprite map}
\usage{
remove_svg_sprite(id, sprite)
}
\arguments{
\item{id}{symbol id of the svg to be removed}

\item{sprite}{\code{xml_document} object with the sprite map}
}
\value{
invisible copy of new sprite map
}
\description{
Remove svg from a sprite map
}
\details{
sprite map is modifyed on place
}
\examples{
library(xml2)
sprite <- xml_new_root("svg", "version" = "1.1", "xmlns" = "http://www.w3.org/2000/svg")
newsvg <- read_xml("<svg version='1.1' xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 68'>
<path d='M55.096 10.462c1.338-1.374.14-3.699-1.322-3.136-1.482.572-1'/>
</svg>")
add_svg_sprite(newsvg, sprite, "new-icon")
remove_svg_sprite("new-icon", sprite)
}
