% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{render_menu_link}
\alias{render_menu_link}
\title{Render menu link}
\usage{
render_menu_link(location, title, active_location = "", icon = NULL)
}
\arguments{
\item{location}{character url with location}

\item{title}{name of the page}

\item{active_location}{name of the active subpage (if matches location then
it gets highlighted), default empty (\code{""})}

\item{icon}{non-mandatory parameter with icon name}
}
\value{
shiny tag link
}
\description{
This function renders horizontal menu item.
}
\examples{
render_menu_link("#subpage1", "SUBPAGE")
}
\seealso{
horizontal_menu
}
