% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_semantic_search.R
\name{render_semantic_search}
\alias{render_semantic_search}
\title{render_semantic_search - wrapper to render shiny semantic dropdown (multiple) search - semantic_search}
\usage{
render_semantic_search(input_name, session, search_query, multiple)
}
\arguments{
\item{input_name}{- string containing input name}

\item{session}{- shiny server session}

\item{search_query}{- function providing a search query}

\item{multiple}{- logical if multiple search (TRUE)}
}
\description{
render_semantic_search - wrapper to render shiny semantic dropdown (multiple) search - semantic_search
}

