% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{right_menu}
\alias{right_menu}
\title{Create Semantic UI Menu on the right}
\usage{
right_menu(...)
}
\arguments{
\item{...}{Menu items to be created in the menu on the right.}
}
\description{
This creates a menu items on the right of a Sematic UI Menu.
}
