% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite.R
\name{rmsvg_sprite}
\alias{rmsvg_sprite}
\title{Remove svg from an sprite map}
\usage{
rmsvg_sprite(id, sprite)
}
\arguments{
\item{id}{symbol id of the svg to be removed}

\item{sprite}{\code{xml_document} object with the sprite map}
}
\value{
invisible copy of new sprite map
}
\description{
Remove svg from an sprite map
}
\details{
sprite map is modifyed on place
}
\examples{
library(xml2)
sprite <- xml_new_root("svg", "version" = "1.1", "xmlns" = "http://www.w3.org/2000/svg")
newsvg <- read_xml("<svg version='1.1' xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 100'> <g> <path d='M50.049,0.3c14.18,0.332,25.969,5.307,35.366,14.923S99.675,36.9,100,51.409c-0.195,11.445-3.415,21.494-9.658,30.146 - yadda yadda yadda'/> </g> </svg>")
addsvg_sprite(newsvg, sprite, "new-icon")
rmsvg_sprite("new-icon", sprite)
}
