% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_field.R
\name{search_field}
\alias{search_field}
\title{Create search field Semantic UI component}
\usage{
search_field(name, search_api_url, default_text = "Search", value = "")
}
\arguments{
\item{name}{Input name. Reactive value is available under input[[name]].}

\item{search_api_url}{Register custom API url with server JSON Response containing fields `title` and `description`.}

\item{default_text}{Text to be visible on serach field when nothing is selected.}

\item{value}{Pass value if you want to initialize selection for search field.}
}
\description{
This creates a default search field using Semantic UI styles with Shiny input. Search field is already initialized
and available under input[[name]]. Search will automatically route to the named API endpoint provided as parameter.
Search is performed across JavaScript searchFields `title` and `description`.
See https://semantic-ui.com/modules/search.html#behaviors for more details.
}

