% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{semantic_DTOutput}
\alias{semantic_DTOutput}
\title{Semantic DT Output}
\usage{
semantic_DTOutput(...)
}
\arguments{
\item{...}{datatable parameters, check \code{?DT::datatable} to learn more.}
}
\value{
DT Output with semanitc style
}
\description{
Semantic DT Output
}
