% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SEMANTIC_PALETTE}
\alias{SEMANTIC_PALETTE}
\title{Semantic colors
https://github.com/Semantic-Org/Semantic-UI/blob/master/src/themes/default/globals/site.variables}
\format{
An object of class \code{character} of length 13.
}
\usage{
SEMANTIC_PALETTE
}
\description{
Semantic colors
https://github.com/Semantic-Org/Semantic-UI/blob/master/src/themes/default/globals/site.variables
}
\keyword{datasets}
