% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_semantic_search.R
\name{semantic_search}
\alias{semantic_search}
\title{semantic_search - function defining the (multiple) search selection dropdown input}
\usage{
semantic_search(name, search_api_url, multiple = FALSE)
}
\arguments{
\item{name}{- string with input name}

\item{multiple}{- logical indicating whether dropdown search is multiple (default FALSE)}

\item{serach_api_url}{- register api url}
}
\description{
semantic_search - function defining the (multiple) search selection dropdown input
}

