% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{semantic_search_choices}
\alias{semantic_search_choices}
\title{semantic_search_choices}
\usage{
semantic_search_choices(name, choices, multiple = FALSE)
}
\arguments{
\item{name}{string with input name}

\item{choices}{vector or a list of choices to search through}

\item{multiple}{logical indicating whether dropdown search is multiple (default FALSE)}
}
\description{
function defining the (multiple) search selection dropdown input using provided choices
}
\examples{
\dontrun{
## server.R example of multiple dropdown search
function(input, output, session) {
choices <- LETTERS
output$letters_results <- shiny::renderUI(semantic_search_choices("serach_result", choices, multiple = TRUE))
}

## ui.R
uiOutput("letters_results")

}
}
