% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{semanticPage}
\alias{semanticPage}
\title{Semantic UI page}
\usage{
semanticPage(..., title = "", theme = NULL)
}
\arguments{
\item{...}{Other arguments to be added as attributes of the main div tag
wrapper (e.g. style, class etc.)}

\item{title}{A title to display in the browser's title bar.}

\item{theme}{Theme name or path}
}
\description{
This creates a Semantic page for use in a Shiny app.
}
\details{
Inside, it uses two crucial options:
- \code{shiny.minified} with a logical value, tells whether it should attach min or full
semnatic css or js (TRUE by default).
- \code{shiny.custom.semantic} if this option has not NULL character \code{semanticPage}
takes dependencies from custom css and js files specified in this path
(NULL by default). Depending on \code{shiny.minified} value the folder should contain
either "min" or standard version.
}
