% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{set_tab_id}
\alias{set_tab_id}
\title{Sets tab id if not provided}
\usage{
set_tab_id(tab)
}
\arguments{
\item{tab}{A tab. Tab is a list of three elements - first
element defines menu item, second element defines tab content, third optional element defines tab id.}
}
\description{
Sets tab id if it wasn't provided
}
