% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.semantic-package.R
\docType{package}
\name{shiny.semantic-package}
\alias{shiny.semantic}
\alias{shiny.semantic-package}
\title{Semantic UI wrapper for Shiny}
\description{
With this library it’s easy to wrap Shiny with Semantic UI
components. Add a few simple lines of code and some CSS classes to give
your UI a fresh, modern and highly interactive look.
}
\section{Options}{

There are a number of global options that affect shiny.semantic as well as
Shiny behavior.The options can be set globally with `options()`
\describe{
\item{shiny.custom.semantic.cdn (defaults to `NULL`)}{This controls from where the css
and javascripts will be downloaded.}
\item{shiny.custom.semantic (defaults to `NULL`)}{This allows to set custom local path
to semantic dependencies.}
\item{shiny.minified (defaults to `TRUE`)}{Defines including JavaScript as a minified or
un-minified file.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://appsilon.github.io/shiny.semantic/}
  \item \url{https://github.com/Appsilon/shiny.semantic}
  \item Report bugs at \url{https://github.com/Appsilon/shiny.semantic/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowicki \email{opensource+kuba@appsilon.com}

Authors:
\itemize{
  \item Filip Stachura \email{filip@appsilon.com}
  \item Dominik Krzeminski
  \item Krystian Igras
  \item Adam Forys
  \item Paweł Przytuła \email{pawel@appsilon.com}
  \item Jakub Chojna \email{jakub.chojna@appsilon.com}
  \item Olga Mierzwa-Sulima \email{olga@appsilon.com}
  \item Tymoteusz Makowski \email{tymoteusz.makowski@appsilon.com}
}

Other contributors:
\itemize{
  \item Ashley Baldry [contributor]
  \item Pedro Manuel Coutinho da Silva \email{pedro@appsilon.com} [contributor]
  \item Kamil Żyła \email{kamil@appsilon.com} [contributor]
  \item Rabii Bouhestine [contributor]
  \item Federico Rivandeira \email{federico@appsilon.com} [contributor]
  \item  Appsilon Sp. z o.o. [copyright holder]
}

}
\keyword{internal}
