% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_input}
\alias{shiny_input}
\title{Universal binding for Shiny input on custom user interface. Using this function one can create various inputs
ranging from text, numerical, date, dropdowns, etc. Value of this input is extracted via jQuery using $().val()
function and exposed as serialized JSON to the Shiny server.}
\usage{
shiny_input(input_id, shiny_ui, value = NULL)
}
\arguments{
\item{input_id}{String with name of this input. Access to this input within server code is normal with input[[input_id]].}

\item{shiny_ui}{UI of HTML component presenting this input to the users. This UI should allow to extract its value with jQuery $().val() function.}

\item{value}{An optional argument with value that should be set for this input. Can be used to store persisten input valus in dynamic UIs.}
}
\description{
Universal binding for Shiny input on custom user interface. Using this function one can create various inputs
ranging from text, numerical, date, dropdowns, etc. Value of this input is extracted via jQuery using $().val()
function and exposed as serialized JSON to the Shiny server.
}

