% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_text_input}
\alias{shiny_text_input}
\title{Create universal Shiny text input binding}
\usage{
shiny_text_input(...)
}
\arguments{
\item{input_id}{String with name of this input. Access to this input within server code is normal with input[[input_id]].}

\item{shiny_ui}{UI of HTML component presenting this input to the users. This UI should allow to extract its value with jQuery $().val() function.}

\item{value}{An optional argument with value that should be set for this input. Can be used to store persisten input valus in dynamic UIs.}
}
\description{
Universal binding for Shiny text input on custom user interface. Value of this input is extracted via jQuery using $().val()
function. This function is just a simple binding over shiny_input. Please take a look at shiny_input documentation for more
information.
}
