% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{show_modal}
\alias{show_modal}
\title{Show Semantic UI modal}
\usage{
show_modal(selector)
}
\arguments{
\item{target}{Javascript selector for the modal that will be displayed.}
}
\description{
This displays a hidden Semantic UI modal.
}
