% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_layout.R
\name{sidebar_panel}
\alias{sidebar_panel}
\title{Creates div containing elements in order or grid container for sidebar panel}
\usage{
sidebar_panel(grid_list, ...)
}
\description{
Creates div containing elements in order or grid container for sidebar panel
}
