% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarLayout.R
\name{sidebarLayout}
\alias{sidebarLayout}
\title{Sidebar Layout ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''}
\usage{
sidebarLayout(sidebar_panel, main_panel, mirrored = FALSE, min_height = "auto")
}
\arguments{
\item{sidebar_panel}{Sidebar panel component}

\item{main_panel}{Main panel component}

\item{mirrored}{If TRUE sidebar is located on the right side,
if FALSE - on the left side (default)}

\item{min_height}{Sidebar layout container keeps the minimum height, if
specified. It should be formatted as a string with css units}
}
\value{
Container with sidebar and main panels
}
\description{
Sidebar Layout ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
}
\examples{
sidebarLayout(
  sidebarPanel("Side Item 1", "Side Item 2", "Side Item 3", width = 20),
  mainPanel("Main 1", "Main 2", "Main 3", "Main 4"),
  mirrored = FALSE,
  min_height = "400px"
)

}
