% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarLayout.R
\name{sidebarPanel}
\alias{sidebarPanel}
\title{Sidebar Panel '''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''}
\usage{
sidebarPanel(..., width = 25)
}
\arguments{
\item{...}{Elements included in sidebar panel}

\item{width}{Width of sidebar panel in percents as numeric value}
}
\description{
Sidebar Panel '''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
}
