% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{simple_checkbox}
\alias{simple_checkbox}
\title{Create Semantic UI checkbox}
\usage{
simple_checkbox(id, label, type = "", is_marked = TRUE, style = NULL)
}
\arguments{
\item{id}{Input name. Reactive value is available under input[[name]].}

\item{label}{Text to be displayed with checkbox.}

\item{type}{Type of checkbox. Please check \code{checkbox_types} for possibilities.}

\item{is_marked}{Defines if checkbox should be marked. Default TRUE.}

\item{style}{Style of the widget.}
}
\description{
This creates a checkbox using Semantic UI styles.
}
\details{
The inputs are updateable by using \code{\link[shiny]{updateCheckboxInput}}.
}
\examples{
simple_checkbox("example", "Check me", is_marked = FALSE)

}
