% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_input.R
\name{slider_input}
\alias{slider_input}
\title{Add Semantic UI slider component}
\usage{
slider_input(name, min, max, value, step = 0.01, n_ticks = 5, color = "")
}
\arguments{
\item{name}{Input name. Reactive value is available under input[[name]].}

\item{min}{Minimum possible value to set.}

\item{max}{Miaksimum possible value to set.}

\item{value}{Initial value.}

\item{step}{Possible step for slider inputs.}

\item{n_ticks}{Specifies how many ticks the slider should display.}

\item{color}{Possible slider color, see: \link{semantic_palette}.}
}
\description{
It implements slider extension provided by: https://github.com/tyleryasaka/semantic-ui-range
}
