% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_input.R
\name{slider_js}
\alias{slider_js}
\title{Java script code that allows passing slider value into shiny}
\usage{
slider_js(slider_id, min, max, init, step, name)
}
\arguments{
\item{slider_id}{Id for div containing the slider.}

\item{min}{Minimum slider value.}

\item{max}{Maximum slider value.}

\item{init}{Initial value set for slider after the page is loaded.}

\item{step}{Step value defining possible values passes into slider.}

\item{name}{Id for input element responsible for passing chosen value into shiny.}
}
\description{
JS code uses range method for slider div element.
The method allows to specify basic parameters for slider such as min, max, etc.
onChange parameter allows to specify callback executed after slider value is changed.
In this case onChange performs the following steps:
\itemize{
  \item{basing on selected value in slider it creates label with selected value and attaches it to slider pointer}
  \item{new value is attached to input element responsible for passing value into Shiny server}
  \item{the value attached to input element replaces the old one that triggers passing it to Shiny server}
}
}
