% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_args}
\alias{split_args}
\title{Split arguments to positional and named}
\usage{
split_args(...)
}
\arguments{
\item{...}{arguments to split}
}
\value{
A list with two named elements:
\itemize{
\item \code{positional}, a list of the positional arguments,
\item \code{named}, a list of the named arguments.
}
}
\description{
Split arguments to positional and named
}
\keyword{internal}
