% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{static_item}
\alias{static_item}
\title{Create Semantic UI Menu Hyperlink Item}
\usage{
static_item(..., item_feature = "", style = "")
}
\arguments{
\item{...}{Other attributes of the menu item such as style and also content of the item: text, icons or labels to be displayed.}

\item{item_feature}{If required, add additional item feature like 'active', 'header', etc.}

\item{style}{Style of the static menu item, e.g. "text-align: center"}
}
\description{
This creates a hyperlink menu item using Semantic UI
}
