% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{steps}
\alias{steps}
\title{Show, Hide or Remove Semantic UI modal}
\usage{
steps(id, steps_list, class = NULL)
}
\arguments{
\item{id}{ID of the Steps that will be displayed.}

\item{steps_list}{A list of steps generated by single_steps.}

\item{class}{(Optional) A character string with the semantic class to be
added to the steps element.}
}
\description{
This displays a hidden Semantic UI modal.
}
\seealso{
single_steps
}
