% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\docType{data}
\name{SUPPORTED_THEMES}
\alias{SUPPORTED_THEMES}
\title{Supported semantic themes}
\format{An object of class \code{character} of length 17.}
\usage{
SUPPORTED_THEMES
}
\description{
Supported semantic themes
}
\keyword{datasets}
