% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPage.R
\name{show_tab}
\alias{show_tab}
\alias{hide_tab}
\title{Show/Hide Tab}
\usage{
show_tab(session = shiny::getDefaultReactiveDomain(), id, target)

hide_tab(session = shiny::getDefaultReactiveDomain(), id, target)
}
\arguments{
\item{target}{}
}
\description{
Dynamically show or hide a \code{\link{tab_panel}} or \code{navbar_menu}
}
