% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPage.R
\name{tabset_panel}
\alias{tabset_panel}
\title{Tabset Panel}
\usage{
tabset_panel(
  ...,
  id = NULL,
  selected = NULL,
  type = c("tabs", "pills", "hidden")
)
}
\description{
Create a tabset that contains \code{\link{tab_panel}}s.
}
