% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{theme_selector}
\alias{theme_selector}
\title{Themes changer}
\usage{
theme_selector()
}
\description{
Themes changer
}
\examples{

if (interactive()) {
library(shiny)
 library(shiny.semantic)
  ui <- semanticPage(
    actionButton("action_button", "Press Me!"),
    textOutput("button_output"),
    theme_selector(),
    textOutput("theme")
  )
  server <- function(input, output, session) {
    output$button_output <- renderText(as.character(input$action_button))
    output$theme <- renderText(as.character(input$theme))
  }
  shinyApp(ui, server)
}

}
