% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\alias{showNotification}
\alias{removeNotification}
\title{Create Semantic UI toast}
\usage{
toast(
  message,
  title = NULL,
  action = NULL,
  duration = 3,
  id = NULL,
  class = "",
  toast_tags = NULL,
  session = shiny::getDefaultReactiveDomain()
)

showNotification(
  ui,
  action = NULL,
  duration = 5,
  closeButton = TRUE,
  id = NULL,
  type = c("default", "message", "warning", "error"),
  session = getDefaultReactiveDomain(),
  ...
)

removeNotification(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID to be added to the toast div. Default "".}

\item{class}{Classes except "ui toast" to be added to the toast. Semantic UI classes can be used. Default "".}

\item{toast_tags}{Other toast elements. Default NULL.}

\item{...}{Content elements to be added to the toast body.
To change attributes of the container please check the `content` argument.}

\item{header}{Content to be displayed in the toast header.
If given in form of a list, HTML attributes for the container can also be changed. Default "".}

\item{content}{Content to be displayed in the toast body.
If given in form of a list, HTML attributes for the container can also be changed. Default NULL.}

\item{footer}{Content to be displayed in the toast footer. Usually for buttons.
If given in form of a list, HTML attributes for the container can also be changed. Default NULL.}

\item{target}{Javascript selector for the element that will open the toast. Default NULL.}

\item{settings}{List of vectors of Semantic UI settings to be added to the toast. Default NULL.}
}
\description{
This creates a toast using Semantic UI styles.
}
\examples{
## Create a simple server toast
library(shiny)
library(shiny.semantic)

ui <- function() {
  shinyUI(
    semanticPage(
      actionButton("show", "Show toast")
    )
  )
}

server = function(input, output) {
  observeEvent(input$show, {
    toast(
      "This is an important message!"
    )
  })
}

}
