% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{toggle_step_state}
\alias{toggle_step_state}
\title{Toggle step state}
\usage{
toggle_step_state(id, state = TRUE, automatic_steps = TRUE, asis = TRUE)
}
\arguments{
\item{id}{ID of step to be toggled}

\item{state}{State of the step, \code{TRUE} stands for enabled}

\item{automatic_steps}{Whether to toggle focus of next step automatically}

\item{asis}{When used inside of Shiny module, \code{TRUE} will disable adding
the namespace to \code{id}}
}
\description{
Toggle step state
}
\seealso{
steps
}
