% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{uibutton}
\alias{uibutton}
\title{Create Semantic UI Button}
\usage{
uibutton(name, label, icon = NULL, type = NULL, ...)
}
\arguments{
\item{name}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link}

\item{icon}{An optional \code{\link{uiicon}()} to appear on the button.}

\item{type}{An optional attribute to be added to the button's class.}

\item{...}{Named attributes to be applied to the button}
}
\description{
Create Semantic UI Button
}
\examples{
uibutton("simple_button", "Press Me!")

}
