% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uicheckbox}
\alias{uicheckbox}
\title{Create Semantic UI checkox}
\usage{
uicheckbox(..., type = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the
tag (e.g. style, childrens etc.)}

\item{type}{Type of checkbox to be used See \code{\link{checkbox_types}} for possible values.}
}
\description{
This creates a checkbox using Semantic UI styles.
}
