% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uidropdown}
\alias{uidropdown}
\title{Create Semantic UI Dropdown}
\usage{
uidropdown(type = "", name, choices = dropdown_menu(), ...,
  in_menu = "FALSE", dropdown_specs = list(), value = NULL)
}
\arguments{
\item{type}{Type of the dropdown. Look at https://semantic-ui.com/modules/dropdown.html for all possibilities.}

\item{name}{Unique name of the created dropdown.}

\item{choices}{Use dropdown_menu to create set of choices, headers and dividers for the dropdown.}

\item{...}{Dropdown items to be created. Use dropdown_item to create new dropdown item.}

\item{in_menu}{TRUE if the dropdown is a menu item. Default is FALSE.}

\item{dropdown_specs}{A list of dropdown functionalities. Look at https://semantic-ui.com/modules/dropdown.html#link3 for all possibilities.}

\item{value}{Default value of the dropdown. Useful when using selection dropdowns with default value selected.}
}
\description{
This creates a dropdown using Semantic UI.
}
