% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uifields}
\alias{uifields}
\title{Create Semantic UI fields tag}
\usage{
uifields(...)
}
\arguments{
\item{...}{Other arguments to be added as attributes of the tag (e.g. style, class or childrens etc.)}
}
\description{
This creates a fields tag using Semantic UI styles.
}
