% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uiform}
\alias{uiform}
\title{Create Semantic UI form tag}
\usage{
uiform(...)
}
\arguments{
\item{...}{Other arguments to be added as attributes of the tag (e.g. style, class or childrens etc.)}
}
\description{
This creates a form tag using Semantic UI styles.
}
