% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uiicon}
\alias{uiicon}
\title{This creates an icon tag using Semantic UI styles.}
\usage{
uiicon(type = "", style = "")
}
\arguments{
\item{type}{A name of an icon. Look at http://semantic-ui.com/elements/icon.html for all possibilities.}

\item{style}{Optional style attribute to add to the tag.}
}
\description{
This creates an icon tag using Semantic UI styles.
}

