% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{uiinput}
\alias{uiinput}
\title{Create Semantic UI Input}
\usage{
uiinput(..., class = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the tag (e.g. style, class or childrens etc.)}

\item{class}{Additional classes to add to html tag.}
}
\description{
This creates an input shell for the actual input
}
\examples{
#' ## Only run examples in interactive R sessions
if (interactive()) {
library(shiny)
library(shiny.semantic)

ui <- semanticPage(
  uiinput(icon("dog"),
          numeric_input("input", value = 0, label = "")
  )
)

server <- function(input, output, session) {
}

shinyApp(ui, server)
}

}
\seealso{
text_input
}
