% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uilist}
\alias{uilist}
\title{Create Semantic UI list with header, description and icons}
\usage{
uilist(content_list, is_divided = FALSE)
}
\arguments{
\item{content_list}{list of lists with fields: `header` and/or `description`,
`icon` containing the list items headers, descriptions (one of these is mandatory)
and icons. Icon column should contain strings with icon names available
here: https://fomantic-ui.com/elements/icon.html}

\item{is_divided}{If TRUE created list elements are divided}
}
\description{
This creates a list with icons using Semantic UI
}
