% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uilist_with_description}
\alias{uilist_with_description}
\title{Create Semantic UI list with header, description and icons}
\usage{
uilist_with_description(data, icon, is_divided = FALSE)
}
\arguments{
\item{data}{A dataframe with two columns `header` and `description` containing the list items
headers and descriptions}

\item{icon}{A string with icon name}

\item{is_divided}{If TRUE created list elements are divided}
}
\description{
This creates a list with icons using Semantic UI
}
\examples{

list_content <- data.frame(
  header = paste("Header", 1:5),
  description = paste("Description", 1:5),
  stringsAsFactors = FALSE
)

# Create a 5 element divided list with alert icons
list_description_icons(list_content, "alert", is_divided = TRUE)
}
