% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uimessage}
\alias{uimessage}
\title{Create Semantic UI Message}
\usage{
uimessage(header, content, type = "", icon, closable = FALSE)
}
\arguments{
\item{header}{Header of the message}

\item{content}{Content of the message. If it is a vector, creates a list of
vector's elements}

\item{type}{Type of the message. Look at
https://semantic-ui.com/collections/message.html for all possibilities.}

\item{icon}{If the message is of the type 'icon', specify the icon.
Look at http://semantic-ui.com/elements/icon.html for all possibilities.}

\item{closable}{Determines whether the message should be closable.
Default is FALSE - not closable}
}
\description{
Create Semantic UI Message
}
