% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uirender.R
\name{uirender}
\alias{uirender}
\title{Render semanticui htmlwidget}
\usage{
uirender(ui, width = NULL, height = NULL, elementId = NULL)
}
\description{
htmlwidget that adds semanticui dependencies and renders in viewer or rmarkdown.
}
