% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{update_numeric_input}
\alias{update_numeric_input}
\alias{updateNumericInput}
\title{Change numeric input value and settings}
\usage{
update_numeric_input(
  session,
  input_id,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL
)

updateNumericInput(
  session,
  inputId,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{min}{Minimum value.}

\item{max}{Maximum value.}

\item{step}{Step size.}

\item{inputId}{the same as \code{input_id}}
}
\description{
Change numeric input value and settings
}
