% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{update_progress}
\alias{update_progress}
\title{Update progress Semantic UI component}
\usage{
update_progress(
  session,
  input_id,
  type = c("increment", "decrement", "label", "value"),
  value = 1
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{input_id}{The id of the input object}

\item{type}{Whether you want to increase the progress bar (\code{"increment"}), decrease the
progress bar (\code{"decrement"}), update the label \code{"label"}, or set it to a specific value (\code{"value"})}

\item{value}{The value to increase/decrease by, or the value to be set to}
}
\description{
Change the value of a \code{\link{progress}} input on the client.
}
