% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{updateActionButton}
\alias{updateActionButton}
\title{Change the label or icon of an action button on the client}
\usage{
updateActionButton(session, inputId, label = NULL, icon = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{icon}{The icon to set for the input object. To remove the current icon, use icon=character(0)}
}
\description{
Change the label or icon of an action button on the client
}
\examples{

if (interactive()){
library(shiny)
library(shiny.semantic)

ui <- semanticPage(
  actionButton("update", "Update other buttons"),
  br(),
  actionButton("goButton", "Go"),
  br(),
  actionButton("goButton2", "Go 2", icon = icon("home")),
  br(),
  actionButton("goButton3", "Go 3")
)

server <- function(input, output, session) {
  observe({
    req(input$update)

    # Updates goButton's label and icon
    updateActionButton(session, "goButton",
                       label = "New label",
                       icon = icon("calendar"))

    # Leaves goButton2's label unchaged and
    # removes its icon
    updateActionButton(session, "goButton2",
                       icon = character(0))

    # Leaves goButton3's icon, if it exists,
    # unchaged and changes its label
    updateActionButton(session, "goButton3",
                       label = "New label 3")
  })
}
shinyApp(ui, server)
}

}
