% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{updateNumericInput}
\alias{updateNumericInput}
\title{Change numeric input value and settings}
\usage{
updateNumericInput(session, inputId, label = NULL, value = NULL,
  min = NULL, max = NULL, step = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{min}{Minimum value.}

\item{max}{Maximum value.}

\item{step}{Step size}
}
\description{
Change numeric input value and settings
}
