% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_value_allowed}
\alias{verify_value_allowed}
\title{Verify that given variable is in the list of allowed values.}
\usage{
verify_value_allowed(variable, values)
}
\arguments{
\item{variable}{variable with a value}

\item{values}{allowed values}
}
\value{
warning if variable not in allowed values
}
\description{
Verify that given variable is in the list of allowed values.
}
