% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_unsupported_args}
\alias{warn_unsupported_args}
\title{Warn that there are not supported arguments}
\usage{
warn_unsupported_args(args)
}
\arguments{
\item{args}{list or vector with extra arguments}
}
\description{
This throws warning if there are parameters not supported by semantic.
}
\keyword{internal}
