% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{with_progress}
\alias{with_progress}
\alias{withProgress}
\alias{setProgress}
\alias{set_progress}
\alias{incProgress}
\alias{inc_progress}
\title{Reporting progress (functional API)}
\usage{
with_progress(
  expr,
  min = 0,
  max = 1,
  value = min + (max - min) * 0.1,
  message = NULL,
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE
)

withProgress(
  expr,
  min = 0,
  max = 1,
  value = min + (max - min) * 0.1,
  message = NULL,
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE,
  ...
)

setProgress(
  value = NULL,
  message = NULL,
  session = getDefaultReactiveDomain(),
  ...
)

set_progress(
  value = NULL,
  message = NULL,
  session = getDefaultReactiveDomain()
)

incProgress(
  amount = 0.1,
  message = NULL,
  session = getDefaultReactiveDomain(),
  ...
)

inc_progress(
  amount = 0.1,
  message = NULL,
  session = getDefaultReactiveDomain(),
  ...
)
}
\arguments{
\item{expr}{The work to be done. This expression should contain calls to
\code{set_progress}.}

\item{min}{The value that represents the starting point of the progress bar.
Must be less tham \code{max}. Default is 0.}

\item{max}{The value that represents the end of the progress bar. Must be
greater than \code{min}. Default is 1.}

\item{value}{Single-element numeric vector; the value at which to set the
progress bar, relative to \code{min} and \code{max}.}

\item{message}{A single-element character vector; the message to be displayed
to the user, or \code{NULL} to hide the current message (if any).}

\item{session}{The Shiny session object, as provided by \code{shinyServer} to
the server function. The default is to automatically find the session by
using the current reactive domain.}

\item{env}{The environment in which \code{expr} should be evaluated.}

\item{quoted}{Whether \code{expr} is a quoted expression (this is not
common).}

\item{...}{Arguments that may have been used in \code{shiny::withProgress}}

\item{amount}{For \code{inc_progress}, the amount to increment the status bar.
Default is 0.1.}
}
\description{
Reports progress to the user during long-running operations.
}
\details{
This package exposes two distinct programming APIs for working with progress.
Using \code{with_progress} with \code{inc_progress} or \code{set_progress}
provide a simple function-based interface, while the \code{\link[=Progress]{Progress()}}
reference class provides an object-oriented API.

Use \code{with_progress} to wrap the scope of your work; doing so will cause a
new progress panel to be created, and it will be displayed the first time
\code{inc_progress} or \code{set_progress} are called. When \code{with_progress}
exits, the corresponding progress panel will be removed.

The \code{inc_progress} function increments the status bar by a specified
amount, whereas the \code{set_progress} function sets it to a specific value,
and can also set the text displayed.

Generally, \code{with_progress}/\code{inc_progress}/\code{set_progress} should
be sufficient; the exception is if the work to be done is asynchronous (this
is not common) or otherwise cannot be encapsulated by a single scope. In that
case, you can use the \code{Progress} reference class.

When migrating from shiny applications, the functions \code{withProgress}, \code{incProgress}
and \code{setProgress} are aliases for \code{with_progress}, \code{inc_progress} and \code{set_progress}.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- semanticPage(
  plotOutput("plot")
)

server <- function(input, output) {
  output$plot <- renderPlot({
    with_progress(message = 'Calculation in progress',
                 detail = 'This may take a while...', value = 0, {
      for (i in 1:15) {
        inc_progress(1/15)
        Sys.sleep(0.25)
      }
    })
    plot(cars)
  })
}

shinyApp(ui, server)
}

}
\seealso{
\code{\link[=Progress]{Progress()}}
}
