% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{actionbutton}
\alias{actionbutton}
\title{Action Button}
\usage{
actionbutton(name, label, icon = NULL, size = "medium", color = NULL,
  ...)
}
\arguments{
\item{name}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link}

\item{icon}{icon type, use \code{uiicon} to insert icon}

\item{size}{character with button size, accepted values: "mini", "tiny", "small"
"medium", "large", "big", "huge", "massive"}

\item{color}{character with semantic color}

\item{...}{Named attributes to be applied to the button}
}
\value{
action button object
}
\description{
Creates an action button whose value is initially zero, and increments by one
each time it is pressed.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
library(shiny)
library(shiny.semantic)
ui <- function() {
  semanticPage(
    shinyUI(
      actionbutton("button", "My button", icon = uiicon("user"),
                   size = "huge", color = "orange")
    )
  )
}
server <- function(input, output) {
  observeEvent(input$button,{
    print("Action")
  })
}
shinyApp(ui, server)
}
}
