% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite.R
\name{addsvg_sprite}
\alias{addsvg_sprite}
\title{Add svg to a sprite map}
\usage{
addsvg_sprite(newsvg, sprite, id = NULL)
}
\arguments{
\item{newsvg}{\code{xml_document} object with svg that will be added}

\item{sprite}{\code{xml_document} object with the sprite map}

\item{id}{sets the id of the new svg inside the sprite map, if is
NULL it looks for the id inside the \code{newsvg} object.}
}
\value{
invisible copy of new sprite map
}
\description{
Add svg to a sprite map
}
\details{
the sprite map object is modified in place
}
\examples{
library(xml2)
sprite <- xml_new_root("svg", "version" = "1.1", "xmlns" = "http://www.w3.org/2000/svg")
newsvg <- read_xml('<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"width="100px" height="100px" viewBox="0 0 100 100" enable-background="new 0 0 100 100" xml:space="preserve"> <g> <path d="M50.049,0.3c14.18,0.332,25.969,5.307,35.366,14.923S99.675,36.9,100,51.409c-0.195,11.445-3.415,21.494-9.658,30.146 - yadda yadda yadda"/> </g> </svg>')
addsvg_sprite(newsvg, sprite, "new-icon")
}
