% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{card}
\alias{card}
\title{Create Semantic UI card tag}
\usage{
card(..., class = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the
tag (e.g. style, class or childrens etc.)}

\item{class}{Additional classes to add to html tag.}
}
\description{
This creates a card tag using Semantic UI styles.
}
