% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_proper_color}
\alias{check_proper_color}
\title{Check if color is set from Semanti-UI palette}
\usage{
check_proper_color(color)
}
\arguments{
\item{color}{character with color name}
}
\value{
Error when \code{color} does not belong to palette
}
\description{
Check if color is set from Semanti-UI palette
}
\examples{
check_proper_color("blue")
}
