% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_shiny_param}
\alias{check_shiny_param}
\title{Checks whether argument included as shiny exclusive parameter}
\usage{
check_shiny_param(name, func, ...)
}
\arguments{
\item{name}{Function argument name}

\item{func}{Name of the function in the}

\item{...}{Arguments passed to the shiny.semantic version of the shiny function}
}
\value{
If the shiny exclusive argument is called in a shiny.semantic, then a message is posted in the UI
}
\description{
A quick function to check a shiny.semantic wrapper of a shiny function to see whether any
extra arguments are called that aren't required for the shiny.semantic version
}
\keyword{internal}
