% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{define_selection_type}
\alias{define_selection_type}
\title{Define search type if multiple}
\usage{
define_selection_type(input_id, multiple)
}
\arguments{
\item{input_id}{character with name}

\item{multiple}{multiple flag}
}
\description{
Define search type if multiple
}
\keyword{internal}
