% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{.onLoad}
\alias{.onLoad}
\title{Internal function that expose javascript bindings to Shiny app.}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
Internal function that expose javascript bindings to Shiny app.
}
