% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{dropdown_menu}
\alias{dropdown_menu}
\title{Create Semantic UI Menu inside dropdown}
\usage{
dropdown_menu(..., type = NULL)
}
\arguments{
\item{...}{Menu items to be created inside dropdown.}

\item{type}{Type of specified menu. It can be: 'left', 'right' or '' (default).}

\item{disable_ui_class}{If TRUE menu is created without ui class - used when UI menu is child of other ui class.}
}
\description{
This creates a menu item that allows to insert more elements inside uidropdown.
}
