% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_random_id}
\alias{generate_random_id}
\title{Generates random id of element}
\usage{
generate_random_id(prefix, id_length = 20)
}
\arguments{
\item{prefix}{character with prefix add to id}

\item{id_length}{numeric with length of id (default 20)}
}
\description{
Generates random id of element
}
