% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{getDeps}
\alias{getDeps}
\title{Add dashboard dependencies to html}
\usage{
getDeps()
}
\value{
Content with appended dependencies.
}
\description{
Internal function that adds dashboard dependencies to html.
}
