% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%:::\%}
\alias{\%:::\%}
\title{::: hack solution to pass CRAN checks}
\usage{
pkg \%:::\% name
}
\arguments{
\item{pkg}{package name}

\item{name}{function name}
}
\value{
function
}
\description{
::: hack solution to pass CRAN checks
}
