% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{header}
\alias{header}
\title{Create Semantic UI header}
\usage{
header(title, description, icon = NULL)
}
\arguments{
\item{title}{Header title}

\item{description}{Subheader text}

\item{icon}{Optional icon name}
}
\description{
This creates a header with optional icon using Semantic UI styles.
}
