% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{label}
\alias{label}
\title{Create HTML label tag}
\usage{
label(...)
}
\arguments{
\item{...}{Other arguments to be added as attributes of the
tag (e.g. style, class or childrens etc.)}
}
\description{
This creates a HTML label tag.
}
