% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{list_element}
\alias{list_element}
\title{Helper function to render list element}
\usage{
list_element(data, is_description, icon, row)
}
\arguments{
\item{icon}{Icon string}

\item{header}{String with header}

\item{description}{String with description}
}
\description{
Helper function to render list element
}
