% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{list_element}
\alias{list_element}
\title{Helper function to render list element}
\usage{
list_element(data, row)
}
\arguments{
\item{data}{data to list; data.frame with required column `header`
and optionally `icon` and/or `description`.}

\item{row}{row character}
}
\description{
Helper function to render list element
}
