% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{list_element}
\alias{list_element}
\title{Helper function to render list element}
\usage{
list_element(data, is_description, is_icon, row)
}
\arguments{
\item{data}{data to list; data.frame with fields
header, icon, description}

\item{is_description}{description flag}

\item{is_icon}{Icon logical to add icon from data}

\item{row}{row character}
}
\description{
Helper function to render list element
}
