% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{list_of_area_tags}
\alias{list_of_area_tags}
\title{Generate list of HTML div elements representing grid areas.}
\usage{
list_of_area_tags(area_names)
}
\arguments{
\item{area_names}{vector with area names}
}
\value{
list of \code{shiny::tags$div}
}
\description{
Generate list of HTML div elements representing grid areas.
}
\details{
This is a helper function used in grid_template()
\preformatted{
  list_of_area_tags(c("header", "main", "footer"))
}
returns the following list:
\preformatted{
  [[1]] <div id="{{ grid_id }}-header" style="grid-area: header; {{ header_custom_css }}">{{ header }}</div>
  [[2]] <div id="{{ grid_id }}-main" style="grid-area: main; {{ main_custom_css }}">{{ main }}</div>
  [[3]] <div id="{{ grid_id }}-footer" style="grid-area: footer; {{ footer_custom_css }}">{{ footer }}</div>
}
}
