% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_layout.R
\name{main_panel}
\alias{main_panel}
\title{Creates div containing elements in order or grid container for main panel}
\usage{
main_panel(grid_list, ...)
}
\arguments{
\item{grid_list}{List to create grid from}

\item{...}{Container's children elements}
}
\description{
Creates div containing elements in order or grid container for main panel
}
