% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{menu_item}
\alias{menu_item}
\title{Create Semantic UI Menu Item}
\usage{
menu_item(..., item_feature = "", style = NULL, href = NULL)
}
\arguments{
\item{...}{Content of the menu item: text, icons or labels to be displayed.}

\item{item_feature}{If required, add additional item feature like 'active', 'header', etc.}

\item{style}{Style of the item, e.g. "text-align: center".}

\item{href}{If NULL (default) menu_item is created with 'div' tag. Otherwise it is created with 'a' tag, and
parameeter defines its href attribute.}
}
\description{
This creates a menu item using Semantic UI
}
\seealso{
menu
}
