% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Create Semantic UI modal}
\usage{
modal(..., id = "", class = "", header = "", content = NULL,
  footer = NULL, target = NULL, settings = NULL, modal_tags = NULL)
}
\arguments{
\item{...}{Content to be displayed in the modal body.}

\item{id}{ID to be added to the modal div. Default "".}

\item{class}{Classes except "ui modal" to be added to the modal. Semantic UI classes can be used. Default "".}

\item{header}{Content to be displayed in the modal header. Default "".}

\item{content}{Content to be displayed in the modal body.}

\item{footer}{Content to be displayed in the modal footer. Usually for buttons. Default NULL.}

\item{target}{Javascript selector for the element that will open the modal. Default NULL.}

\item{settings}{List of vectors of Semantic UI settings to be added to the modal. Default NULL.}

\item{modal_tags}{Other modal elements. Default NULL.}
}
\description{
This creates a modal using Semantic UI styles.
}
