% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{panel}
\alias{panel}
\title{Creates div containing children elements in separate rows}
\usage{
panel(...)
}
\arguments{
\item{...}{Container's children elements}
}
\value{
Div containing children elements in separate rows
}
\description{
Creates div containing children elements in separate rows
}
