% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{panel}
\alias{panel}
\title{Creates div containing elements in order or grid container for specific panels}
\usage{
panel(args)
}
\arguments{
\item{grid_list}{List to create grid from}

\item{...}{Container's children elements}
}
\value{
Div containing elements or grid container for specific panel
}
\description{
Creates div containing elements in order or grid container for specific panels
}
