% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_semantic_search.R
\name{register_search}
\alias{register_search}
\title{register_search}
\usage{
register_search(session, search_query)
}
\arguments{
\item{session}{- shiny server session}

\item{search_query}{- function providing a search query}
}
\description{
register_search
}

