% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{semantic_search_api}
\alias{semantic_search_api}
\title{semantic_search_api}
\usage{
semantic_search_api(name, search_api_url, multiple = FALSE)
}
\arguments{
\item{name}{string with input name}

\item{search_api_url}{register api url}

\item{multiple}{logical indicating whether dropdown search is multiple (default FALSE)}
}
\description{
function defining the (multiple) search selection dropdown input using REST API
}
