% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{semanticPage}
\alias{semanticPage}
\title{Semantic UI page}
\usage{
semanticPage(title = "", ...)
}
\arguments{
\item{title}{A title to display in the browser's title bar.}
}
\description{
This creates a Semantic page for use in a Shiny app.
}

