% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticPage.R
\name{semanticPage}
\alias{semanticPage}
\title{Semantic UI page}
\usage{
semanticPage(..., title = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the main div tag wrapper (e.g. style, class etc.)}

\item{title}{A title to display in the browser's title bar.}
}
\description{
This creates a Semantic page for use in a Shiny app.
}
