% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{package}
\name{shiny.semantic}
\alias{shiny.semantic}
\alias{shiny.semantic-package}
\title{Semantic UI wrapper for Shiny}
\description{
With this library it’s easy to wrap Shiny with Semantic UI
components. Add a few simple lines of code and some CSS classes to give
your UI a fresh, modern and highly interactive look.
}
\section{Options}{

There are a number of global options that affect shiny.semantic as well as
Shiny behavior.The options can be set globally with `options()`
\describe{
\item{shiny.custom.semantic.cdn (defaults to `NULL`)}{This controls from where the css
and javascripts will be downloaded.}
\item{shiny.custom.semantic (defaults to `NULL`)}{This allows to set custom local path
to semantic dependencies.}
\item{shiny.minified (defaults to `TRUE`)}{Defines including JavaScript as a minified or
un-minified file.}
}
}

