% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{show_modal}
\alias{show_modal}
\alias{remove_modal}
\alias{remove_all_modals}
\alias{removeModal}
\alias{hide_modal}
\title{Show, Hide or Remove Semantic UI modal}
\usage{
show_modal(id, session = shiny::getDefaultReactiveDomain(), asis = TRUE)

remove_modal(id, session = shiny::getDefaultReactiveDomain(), asis = TRUE)

remove_all_modals(session = shiny::getDefaultReactiveDomain())

removeModal(session = shiny::getDefaultReactiveDomain())

hide_modal(id, session = shiny::getDefaultReactiveDomain(), asis = TRUE)
}
\arguments{
\item{id}{ID of the modal that will be displayed.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{asis}{A boolean indicating if the id must be handled as is (TRUE) or
will be it must be namespaced (FALSE)}
}
\description{
This displays a hidden Semantic UI modal.
}
\seealso{
modal
}
