% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{show_modal}
\alias{show_modal}
\alias{remove_modal}
\alias{hide_modal}
\title{Show, Hide or Remove Semantic UI modal}
\usage{
show_modal(id, session = shiny::getDefaultReactiveDomain())

remove_modal(id, session = shiny::getDefaultReactiveDomain())

hide_modal(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of the modal that will be displayed.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
This displays a hidden Semantic UI modal.
}
