% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_layout.R
\name{sidebar_layout}
\alias{sidebar_layout}
\title{Sidebar layout composed of main and sidebar panels}
\usage{
sidebar_layout(text)
}
\arguments{
\item{text}{Some text}
}
\value{
some test function
}
\description{
Sidebar layout composed of main and sidebar panels
}
\examples{
sidebar_layout("Kuba")

}
