% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPage.R
\name{tab_panel}
\alias{tab_panel}
\title{Tab Panel}
\usage{
tab_panel(title, ..., value = title, icon = NULL)
}
\arguments{
\item{title}{Display title for tab}

\item{...}{UI elements to include within the tab}

\item{value}{The value that should be sent when \code{\link{tabset_panel}} reports that this tab is selected.
If omitted and \code{\link{tabset_panel}} has an id, then the title will be used.}

\item{icon}{Optional icon to appear on the tab.
This attribute is only valid when using a \code{tab_panel} within a \code{\link{navbar_page}}.}
}
\value{
A tab that can be passed to \code{\link{tabset_panel}}.
}
\description{
Create a tab panel
}
\examples{
tabset_panel(
  tab_panel("Plot", shiny::plotOutput("plot")),
  tab_panel("Summary", shiny::verbatimTextOutput("summary")),
  tab_panel("Table", shiny::tableOutput("table"))
)

}
\seealso{
\code{\link{tabset_panel}}
}
