% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{uicalendar}
\alias{uicalendar}
\title{Create Semantic UI Calendar}
\usage{
uicalendar(
  name,
  value = NULL,
  placeholder = NULL,
  type = "date",
  min = NA,
  max = NA
)
}
\arguments{
\item{name}{Input name. Reactive value is available under \code{input[[name]]}.}

\item{value}{Initial value of the numeric input.}

\item{placeholder}{Text visible in the input when nothing is inputted.}

\item{type}{Select from \code{'year'}, \code{'month'}, \code{'date'} and \code{'time'}}

\item{min}{Minimum allowed value.}

\item{max}{Maximum allowed value.}
}
\description{
This creates a default calendar input using Semantic UI. The input is available
under \code{input[[name]]}.
}
\examples{
# Basic calendar
uicalendar(name = "date_start")

# Calendar with max and min
uicalendar(
  name = "date_finish",
  placeholder = "Select End Date",
  min = "2019-01-01",
  max = "2020-01-01"
)

# Selecting month
ui_calendar(
  name = "month",
  type = "month"
)
}
