% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{uicheckbox}
\alias{uicheckbox}
\title{Create Semantic UI checkox}
\usage{
uicheckbox(..., type = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the
tag (e.g. style, childrens etc.)}

\item{type}{Type of checkbox to be used.

The following \code{type}s are allowed:
\itemize{
\item{NULL}{The standard checkbox (default)}
\item{toggle}{Checkbox with toggle form}
\item{slider}{Checkbox with slider form}
}}
}
\description{
This creates a checkbox using Semantic UI styles.
}
