% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{uidropdown}
\alias{uidropdown}
\title{Create Semantic UI Dropdown}
\usage{
uidropdown(..., type = "", name, is_menu_item = FALSE,
  dropdown_specs = list())
}
\arguments{
\item{...}{Dropdown content.}

\item{type}{Type of the dropdown. Look at https://semantic-ui.com/modules/dropdown.html for all possibilities.}

\item{name}{Unique name of the created dropdown.}

\item{is_menu_item}{TRUE if the dropdown is a menu item. Default is FALSE.}

\item{dropdown_specs}{A list of dropdown functionalities. Look at https://semantic-ui.com/modules/dropdown.html#/settings for all possibilities.}

\item{choices}{Use dropdown_menu to create set of choices, headers and dividers for the dropdown.}
}
\description{
This creates a dropdown using Semantic UI.
}
\examples{

uidropdown(
  "Dropdown menu",
  uiicon(type = "dropdown"),
  dropdown_menu(
    menu_header("Header"),
    menu_divider(),
    menu_item("Option 1"),
    menu_item("Option 2")
  ),
  name = "dropdown_menu",
  dropdown_specs = list("duration: 500")
)

}
