% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{uiinput}
\alias{uiinput}
\title{Create Semantic UI Input}
\usage{
uiinput(..., class = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the tag (e.g. style, class or childrens etc.)}

\item{class}{Additional classes to add to html tag.}
}
\description{
This creates an input shell for the actual input
}
\examples{
library(shiny)
library(shiny.semantic)

# Text input
uiinput(
  tags$label("Text input"),
  text_input("ex", type = "text", placeholder = "Enter Text")
)

}
\seealso{
text_input
}
