% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uirender.R
\name{uirender}
\alias{uirender}
\title{Render semanticui htmlwidget}
\usage{
uirender(ui, width = NULL, height = NULL, element_id = NULL)
}
\arguments{
\item{ui}{UI, which will be wrapped in an htmlwidget.}

\item{width}{Fixed width for widget (in css units). The default is NULL,
which results in intelligent automatic sizing.}

\item{height}{Fixed height for widget (in css units). The default is NULL,
which results in intelligent automatic sizing.}

\item{element_id}{Use an explicit element ID for the widget
(rather than an automatically generated one).}
}
\description{
htmlwidget that adds semanticui dependencies and
renders in viewer or rmarkdown.
}
\examples{
library(shiny)
library(shiny.semantic)
uirender(
  card(
    div(
      class="content",
      div(class="header", "Elliot Fu"),
      div(class="meta", "Friend"),
      div(class="description", "Elliot Fu is a film-maker from New York.")
    )
  )
)

}
