% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{update_rating_input}
\alias{update_rating_input}
\title{Update rating}
\usage{
update_rating_input(session, input_id, label = NULL, value = NULL)
}
\arguments{
\item{session}{shiny object with session info}

\item{input_id}{rating input name}

\item{label}{character with updated label}

\item{value}{new rating value}
}
\description{
Change the value of a rating input on the client. Check
\code{rating_input} to learn more.
}
