% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{update_rating}
\alias{update_rating}
\alias{updateRating}
\title{Update rating}
\usage{
update_rating(session, input_id, label = NULL, value = NULL)

updateRating(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{shiny object with session info}

\item{input_id}{rating input name}

\item{label}{character with updated label}

\item{value}{new rating value}
}
\description{
check \code{rating} to learn more.
}
