% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{updateActionButton}
\alias{updateActionButton}
\title{Change the label or icon of an action button on the client}
\usage{
updateActionButton(session, inputId, label = NULL, icon = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{icon}{The icon to set for the input object. To remove the current icon, use icon=character(0)}
}
\description{
Change the label or icon of an action button on the client
}
